Add Story tests to integrtion phase in Application module
A 'story' test is a test definition that test certain scenarios that
can run against multiple environment. Normally created close to a
module level with a mocked/doubled backend for development test and
speed, but the same story can also be executed as part of the
verification phase of the final deployment.

Issue: #40