Calculate the Base path dynamically
Change index.html to index.jsp so we can pass a request
variable from the AppFilter to the index page that describes the
BASE path for this deplomyent.

BASE is different depending on if it's deployed as a named deployment
or as a ROOT deployment.

BASE is used by Angular to calculate the $location.paths.