Add Security to StateChanging REST methods
* POST, PUT, DELETE now require authentication
* Change to use Twitter Sign In, instead of app authentication
* Workaround AGOVA-53

When the Security Module is included in the deployment
all REST services are locked down. This means any state changing
methods will require authentication. The Security Module test-jar
comes with a Arquillian Extension that will setup do a full
authentication in AfterDeploy and setup RestAssured with a
valid Http Session. This allow the conference, user, attachment
REST Stories used as final functional tests in the Application
to work with our without the Security Module included. The REST
tests have no knowledge of the Security Module.

Issue: #52