Update Attachment Domain to rely on an immutable data structure
* Remove custom serialize/deserialize

The data structure is immutable to avoid mutations of the
Object it self when doing operations within a TX on a Local
cache.

Even tho the tx is rolledback and the new instance is not stored
in cache, Infinispan will return the same instance of the object.
When the Object is mutable the rollbacked state will be available
outside the tx. Using a immutable Object causes Infinispan to
never store the new object on put when rollbacked, so the old one
is returned.

Based on comments by @Sanne